refactor(core): 🔨 split EventObject into partial classes
